requirements.txt
aiosqlite==0.19.0
discord.py==2.2.2
watchfiles==0.19.0
aiofiles==23.1.0
aiopath==0.6.11
beautifulsoup4==4.12.1
pytimeparse==1.1.8
